more size_t instead of int

+++ examples/mnist/src/processData.hpp
@@ -6,5 +6,5 @@ using namespace mygrad;
 
 Tensor loadMnistImages(const std::string& path);
 Tensor loadMnistLabels(const std::string& path); 
-void visualizeImage(const Tensor& images, const Tensor& labels, int index); 
-void visualizeImage(const Tensor& images, int index);
+void visualizeImage(const Tensor& images, const Tensor& labels, size_t index); 
+void visualizeImage(const Tensor& images, size_t index);
